# apps/command/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KSP Command OS — Read API",
    "version": "1.0.0",
    "description": "RLS-scoped access to KSP Command OS for AI connectors (ChatGPT Actions / Claude MCP). Every request runs as the authenticated user; it never returns or changes more than that user could in the app. Writes are limited to low-risk internal actions (create task, add comment, create mission); sensitive/material actions (finance, access grants, deploys, publishing to clients) are intentionally not exposed."
  },
  "servers": [{ "url": "https://REPLACE-WITH-YOUR-COMMAND-HOST/api/v1", "description": "Command app v1 API (replace host)" }],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/proposals": {
      "post": {
        "operationId": "proposeForApproval",
        "summary": "Propose a sensitive/material action for HUMAN approval. Does not execute — files an approval request a person decides in the Decisions module.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": {
            "type": "object", "required": ["approvalType", "riskLevel"],
            "properties": {
              "approvalType": { "type": "string", "enum": ["executive_access", "bank_destination", "high_value_payment", "contract_change", "pricing_exception", "period_reopen", "bulk_export", "production_credential", "rls_auth_change", "protected_deletion", "agent_autonomy", "high_risk_publication", "deployment_exception"] },
              "riskLevel": { "type": "string", "enum": ["low", "medium", "high", "critical"] },
              "amountMinor": { "type": "integer", "minimum": 0 },
              "dueAt": { "type": "string", "description": "Optional RFC3339/date." }
            }
          } } }
        },
        "responses": { "201": { "description": "Filed for human approval.", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "id": { "type": "string" }, "status": { "type": "string" } } } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "A Supabase user access token. The request runs as that user with their RLS in force." }
    },
    "responses": {
      "Unauthorized": { "description": "Missing/invalid token, or the user has no active internal membership.", "content": { "application/json": { "schema": { "type": "object", "properties": { "error": { "type": "string" } } } } } },
      "BadRequest": { "description": "Invalid input.", "content": { "application/json": { "schema": { "type": "object", "properties": { "error": { "type": "string" } } } } } },
...
    },
...
  }
}